feat(suite-native): transaction detail overview section
- edit styles of transaction detail overview section by using atoms. With new design, some unnecessary styles could now be deleted and are replaced with simplified version

# Conflicts:
#	suite-native/intl/src/messages.ts
